fix(use-i18n): do not use typescript global namespace
