--- types/global.d.ts
@@ -1,5 +1,0 @@
-export {};
-
-declare global {
-  type PrimitiveType = string | number | boolean | null | undefined | Date;
-}
